test(io): move reloading related tests to test_reloading.py
